app/requirements.txt
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-environ==0.11.2
celery==5.3.6
djangorestframework-simplejwt==5.3.1
pandas==2.2.2
openpyxl==3.1.2
xlrd==2.0.1